Improve build scripts
Splits the monaco-editor script into separate ones.
Reuses the relaxation of peer dependency versions for the language-server.
