+++ libs/monaco-editor/project.json
@@ -44,7 +44,9 @@
       "dependsOn": ["build"],
       "options": {
         "commands": [
-          "node tools/scripts/monaco-editor/fix-peer-dependencies.mjs @jvalue/monaco-editor"
+          "node tools/scripts/relax-peer-dependency-versions.mjs @jvalue/monaco-editor",
+          "node tools/scripts/monaco-editor/delete-vscode-peer-dependency.mjs @jvalue/monaco-editor",
+          "node tools/scripts/monaco-editor/relax-react-version.mjs @jvalue/monaco-editor"
         ]
       }
     },
